refactor(common): 优化集合初始容量和常量定义
- 统一多处代码中集合的初始容量为常量 COLLECTION_CAPACITY，提升性能和可维护性
- 将多个配置类中硬编码的默认值提取为私有静态常量，便于统一管理和修改
- 使用 BigDecimal 替代 double 进行阈值和置信度相关的数字计算，提高精度
- 调整部分导入顺序和去除无用空行，改善代码整洁度
- 统一默认参数定义，涵盖超时、Token限制、缓存TTL等多个子模块
- 优化条件判断表达式，避免隐式类型转换风险

+++ ydsz-common/ydsz-common-audit/src/main/java/com/njydsz/common/audit/core/AuditMetricsBinder.java
@@ -81,7 +81,7 @@ public class AuditMetricsBinder implements MeterBinder {
         auditRecorder,
         recorder -> {
           if (recorder instanceof AsyncAuditRecorder asyncRecorder) {
-            return asyncRecorder.getQueueUsageRatio();
+            return asyncRecorder.getQueueUsageRatio().doubleValue();
           }
           return 0.0;
         });
