+++ ydsz-common/ydsz-common-docs/src/main/java/com/njydsz/common/docs/security/pii/impl/IdCardDetector.java
@@ -1,5 +1,6 @@
 package com.njydsz.common.docs.security.pii.impl;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -55,7 +56,7 @@ public class IdCardDetector implements PiiDetector {
               .maskedValue(mask(match.rawValue()))
               .startIndex(match.startIndex())
               .endIndex(match.endIndex())
-              .confidence(0.98)
+              .confidence(BigDecimal.valueOf(0.98))
               .build());
     }
